Phase 4 Wave 2.1: real overlay region map + LevelId coverage invariant
- practice_overlay.c: LevelId -> SceneId -> ovl_i1..6 map (17 saveables);
  exclusion list for non-saveable enums; lazy gDmaTable index by ROM start;
  CRC32-IEEE build-id over in-RAM overlay image with active-overlay guard.
- practice_overlay_request_load remains stub until Phase 5.
- tools/practice_invariants.py: check_overlay_table_complete().
- lib/crc32: use uint32_t len to avoid stddef.h conflict in IDO+N64 build.

Made-with: Cursor

+++ lib/crc32.c
@@ -7,9 +7,9 @@ uint32_t crc32_init(void) {
     return CRC32_INIT;
 }
 
-uint32_t crc32_update(uint32_t state, const void *data, size_t len) {
+uint32_t crc32_update(uint32_t state, const void *data, uint32_t len) {
     const uint8_t *p = (const uint8_t *)data;
-    size_t i;
+    uint32_t i;
     int bit;
 
     if (len == 0) {
@@ -33,6 +33,6 @@ uint32_t crc32_finalize(uint32_t state) {
     return state ^ 0xFFFFFFFFu;
 }
 
-uint32_t crc32(const void *data, size_t len) {
+uint32_t crc32(const void *data, uint32_t len) {
     return crc32_finalize(crc32_update(crc32_init(), data, len));
 }
